requirements.txt
flask
python-dotenv
requests
beautifulsoup4
gunicorn
flask-compress
python-frontmatter
markdown
google-cloud-secret-manager>=2.16.0